YT-21906: Merge to 24.1 0ea1387df6c5adc54a4eef06ab2908c6728b4b50

+++ yt/yt/server/lib/cellar_agent/occupant.h
@@ -79,7 +79,7 @@ struct ICellarOccupant
     virtual void PopulateAlerts(std::vector<TError>* alerts) const = 0;
 
     // Auxiliary.
-    virtual int EstimateChangelogMediumBytes(int payload) const = 0;
+    virtual i64 EstimateChangelogMediumBytes(i64 payload) const = 0;
 };
 
 DEFINE_REFCOUNTED_TYPE(ICellarOccupant)
